requirements.txt
numpy==1.24.3
Flask==2.3.3
pandas==2.1.0
Werkzeug==2.3.7
openpyxl==3.1.2
xlrd==2.0.1
pystray==0.19.4
Pillow==10.0.0
Flask-Session==0.5.0
reportlab==3.6.12
pdfkit==1.0.0
PyMuPDF==1.22.5
selenium==4.11.2
fpdf==1.7.2
xhtml2pdf==0.2.11
imgkit==1.2.3
opencv-python==4.8.0.76
gunicorn==21.2.0